// src/interface.d.ts
// Colors
export type Color = 'primary' | 'secondary' | 'danger' | 'warning' | 'success' | 'danger';

export type PredefinedGlobalSize = 'small' | 'medium' | 'large' | 'big';

// Buttons
export type PredefinedButtonType = 'button' | 'submit' | 'reset';
export type ButtonType = 'button' | 'submit' | 'reset';